minor changes in ArticlesScreen
