+++ app/src/main/java/srimani7/apps/feedfly/audio/SongState.kt
@@ -1,0 +1,6 @@
+package srimani7.apps.feedfly.audio
+
+data class SongState(
+    val duration: Long = 0,
+    val isPlaying: Boolean = false
+)
